+++ EDDI/ChangeLog.md
@@ -1,4 +1,12 @@
-# CHANGE LOG
+﻿# CHANGE LOG
+
+### 2.4.3
+  * Speech Responder
+    * Add Emphasize() Cottle function. Details on how to use this are in the SpeechResponder documentation
+    * Add SpeechPitch() Cottle function. Details on how to use this are in the SpeechResponder documentation
+    * Add SpeechRate() Cottle function. Details on how to use this are in the SpeechResponder documentation
+    * Add SpeechVolume() Cottle function. Details on how to use this are in the SpeechResponder documentation
+    * Fixed a bug that was preventing the Play() function from working properly
 
 ### 2.4.3
   * Core